Split 'lsvast' source code into separate files
As the tool continues to grow, it becomes more and more
untenable to keep all of the source code in a single file.

This commit should contains no changes in logic, just a
reorganization of the existing source files.
