+++ tools/lsvast/CMakeLists.txt
@@ -6,7 +6,8 @@ if (NOT VAST_ENABLE_LSVAST)
   return()
 endif ()
 
-add_executable(lsvast lsvast.cpp)
+add_executable(lsvast src/lsvast.cpp src/print_index.cpp
+                      src/print_partition.cpp src/print_segment.cpp)
 target_link_libraries(lsvast PRIVATE vast::libvast vast::internal)
 install(TARGETS lsvast DESTINATION "${CMAKE_INSTALL_BINDIR}")
